[Quantization] Add metadata porting for nodes added by quantization
Summary:
This diff adds adding metadata to q-dq nodes by inferring the
quatization intent from node annotations. Annotations on the node are
way for user to specify how a node or subgraph is supposed to be
quantized. We continue to use that information to copy metadata on Q/DQ
node from appropriate nodes.

Test Plan:

Reviewers:

Subscribers:

Tasks:

Tags:

ghstack-source-id: e0aaa2243d3d8deb285626f6a6491b8bde85ad0a
Pull Request resolved: https://github.com/pytorch/pytorch/pull/107107